Fix typo in CRC16 return val

+++ src/osdp_common.c
@@ -89,7 +89,7 @@ uint16_t crc16_itu_t(uint16_t seed, const uint8_t * src, size_t len)
 }
 #endif
 
-uint8_t compute_crc16(const uint8_t *buf, size_t len)
+uint16_t compute_crc16(const uint8_t *buf, size_t len)
 {
 	return crc16_itu_t(0x1D0F, buf + 1, len - 1);
 }
